[reduce_delta][media][ums][vtg] Partial revert of video texture black frame creation.
:Release Notes:
Partial revert of video texture black frame creation.

:Detailed Notes:
Partially reverted patch review:166485. It was changing the way the
black frame was created for video playing. But it was not actually needed with new
upstream format provided for black frame.

:Testing Performed:
Built and tested on TV.
Test case:
1. Open browser on TV
2. Open http://media.w3.org/2010/05/video/movie_300.mp4

The video works with and without the patch.

This reduces 5 lines of delta with upstream.

:QA Notes:

:Issues Addressed:
[PLAT-24903] Media not visible for video URL
[PLAT-28078] Migrate [vtg] patches
[PLAT-40026] WebAppMgr crashes at location /usr/lib/libcbe.so(_ZNK5media10VideoFrame4dataEj+0x1)
[PLAT-50746] Upstream delta analysis: src/media
[PLAT-51205] Analysis of delta in media::VideoFrame::CreateBlackFrame

Change-Id: Id2ea965eac28b1f725270bf73360abb73110e136