[reduce_delta][icu] Rework ICU build defines to reduce upstream files delta.
:Release Notes:
Rework ICU build defines so downstream does not need a change in upstream code.

:Detailed Notes:
Upstream GYP scripts had a bug in the way ICU_UTIL_DATA_IMPL value
was passed, so it would only be defined in the modules directly
using ICU.

But it was also used in services/shell. As the defines were missing,
the code would choose the wrong code path. A build fix was added to
take directly the configuration variable and enable the right implementation.
But it is better to just provide from LG build scripts the right defines
so we do not need an extra code change.

This reduces delta on upstream files, but increases slightly the delta
on downstream files. This change (and the previous fix) will not be needed
once we move to GN as GN script properly declares the defines for all
modules using them.

:Testing Performed:
Built on device and desktop.

:QA Notes:

:Issues Addressed:
[PLAT-51367] Upstream delta analysis: misc folders
[PLAT-52149] chromium53: reduce trivial delta in src/ash, src/sandbox, src/services

Change-Id: I1509dc880e7c48d8fb49fda971a4ce84f4ef51ca