[reduce_delta][sandbox] Remove unneeded guards
:Release Notes:
Remove sandbox unneeded guards for pclinux build.

:Detailed Notes:
There was a change for building PCLinux Ubuntu 14.04 build of sandbox
code, that does not seem to be required nowadays (even in Ubuntu 14.04).
So remove the changes.

It was not clear in the original change why it was required. It was
disabling code for protecting Clone operations.

:Testing Performed:
Build on PCLinux on Ubuntu 14.04, device

:QA Notes:

:Issues Addressed:
[PLAT-51367] Upstream delta analysis: misc folders
[PLAT-52149] chromium53: reduce trivial delta in src/ash, src/sandbox, src/services

Change-Id: Id4000d0f72b8e3b10c6dd3d70a2946f6c7f60690