[ozone] Add stub wl_output geometry event
:Release Notes:
Add stub implementation for wl_output geometry event.

:Detailed Notes:
LG AGL wayland compositor is sending geometry events through
wl_output, but is missing the handler. That makes chrome launch
fail.

Add a stub for it. Ideally a full implementation should be
provided to react on the fly to device scale changes.

:Testing Performed:
It does not crash with geometry events now.

:QA Notes:

:Issues Addressed:

Change-Id: I0e72f639bdd179cbadb7d4a6326b43e841174b1d
Tested-by: Jose Dapena <jose.dapena@lge.com>
Reviewed-by: Jani Hautakangas <jani.hautakangas@lge.com>
Reviewed-by: Jose Dapena <jose.dapena@lge.com>