[webos_webview][ozone][agl] Add API for setting the surface ID of web app windows.
:Release Notes:
Add a way to set the surface ID for web app windows.

:Detailed Notes:
In AGL, applications are expected to notify their PID to wayland, so it
matches the one of the executable created from AFM. But in WAM, this
PID will not be the one of browser process.

So this change adds an API so WAM can set the PID on launch.

:Testing Performed:

:QA Notes:

:Issues Addressed:

Change-Id: I898b54adff06e17bba44c07e30a6b5024296c3d8