[media][ums][rpi] Add UMS Internal API version 2 in umediaclient_impl
:Release Notes:
Add UMS Internal API version 2 in umediaclient_impl

:Detailed Notes:
we need to modify umediaclient_impl in chromium53 such that we use
the new API (UMS Internal API version 2) for reference platform.
but other platform should continue to use the older API.
In addition, we need to add video and audio callbacks to support
adaptive streaming. when video changed during adaptive streaming,
we can get the changed video information.
Also we need to add a couple of work in webmediaplayer_ums.

:Testing Performed:

:QA Notes:
Depends on UMS changes in submissions/184.open.10
519608 Separate UMS_INTERNAL_API_VERSIONS for media information notification

:Issues Addressed:
[PLAT-62463] CCC :
[PLAT-61608] [RPi] Resolution change issue in adaptive streaming of
                   uri player

Change-Id: I083a4521618ba187bdbb982e821006d20548194a