[ozone] Abort after retries if WAM can't initialize the display
:Release Notes:
Abort after retries if WAM can't initialize the display

:Detailed Notes:
Previously WAM would fail to establish a connection to wayland
server and not abort, so it would hang around and segfault
whenever a new renderer was spawn and tried to use the display.

LSM needs some time to initialize and setup the wayland sockets
with the right permissions before WAM can succeed establishing
their connection. But systemd launches WAM right away after
spawing LSM.

So implement a retry mechanism in WAM that makes the race
triggering very unlikely, so WAM insists on establishing the
connection for 10 times in 5 seconds. After that it really
aborts (as opposed to previous behavior) and gives a return
error code back to systemd.

This also will not propagate the error upward in the stack
because it's a fatal condition that can't be properly handled.

:Testing Performed:
Tested locally wam keeps retrying for 10 times in 5 seconds
until it gives up and returns an error to systemd.

:QA Notes:

:Issues Addressed:
[PLAT-62975] WAM crash after restart LSM
[PLAT-63635] Port more Blink changes from chromium53 + wam master to
             @open branches

Change-Id: I4635a723238d2445fbae4983f96a80da031c13aa