[native_scroll] Ignore animation having WaitingToCancelOnCompositor
:Release Notes:
Ignore animation having WaitingToCancelOnCompositor state.

:Detailed Notes:
Remaining animation of CompositorScroll could be fired after
ProgrammaticScroll event is fired.
The animation could not be canceled immediately because the creation
and the animation of scroll are processed in different threads.

:Testing Performed:
Built and tested below scenario.

1. Load enact VirtualListNative test
2. With arrow keys focus "test 6" entry.
3. Move focus to bottom.
4. Press 'down' key and press 'page down' key while red focus box
   is moving down.
5. Focus disappear.

Expected Result: Focus moves down normally and does not disappear.

:QA Notes:

:Issues Addressed:
[PLAT-62467] VirtualListNative: focus disappears and unable to recover (1 QE)
[PLAT-63635] Port more Blink changes from chromium53 + wam master to
             @open branches

Change-Id: Ia5fcedd427c1508ed51964fcaa53a7fe2c89fe09