[ose][mse][ums]Fix the bug on loss of the lower part of video
:Release Notes:
Bug fix on loss of the lower part of video.

:Detailed Notes:
This commit is for bug fix about loss of lower part of video is on Enact
browser of webosose.

:Testing Performed:
Tests should be performed on all the followings
- Enact browser
1. Open the Youtube and Naver TV
- tv.naver.com (UMS)
- www.youtube.com (MSE)
2. Play video and check if the video contents is in the video frame
correctly

Expected Result: there is no loss of lower part of video in the video
frame.

:QA Notes:
:Issues Addressed:
[PLAT-61711] [webOS OSE] Video window geometry is incorrect

Change-Id: I6366f792b3089a968f68e1827e176fae88033642