[open][media][gmp]Manage media-pipeline status on its visibility
:Release Notes:
Manage media pipeline status based on its visibility

:Detailed Notes:
To draw multiple videos correctly with using punch-hole, z-order
of each video should be synced with the video layer of chromium
compositor when media-pipeline draws the video. But z-order is decided
by the order of creating of media-pipeline resource and it is not
configurable from the chromium side in our platform currently.
So if media resource for Ad created later than for main video, Ad video
has higher z-order than main video and they have overlaped position so
Ad video was drawn on top of the main video. But when main video comes
up on the screen, actually video element of Ad doesn't have a layout
position on the screen. So if we are possible to let media-pipeline know
the information of the visibility of the video, this issue can be fixed
without correct the z-order of videos.
In this patch, we detect a video visibility based on its layer position
and then let media-pipeline know its visibility to solve this problem.

:Testing Performed:
Tested below issue in RPI target.

:QA Notes:

:Issues Addressed:
[PLAT-61713] [webOS OSE] When we switch from ad. to main video content,
for example on youtube, we cannot make transition from ad. to main video
content seamlessly
[PLAT-64928] [webOSOSE] When you play any YouTube video, Ads overlap.

Change-Id: If1a91eae897eb777419ca5b7e4a41d22a0d0b463
Reviewed-on: http://gpro.lgsvl.com/214969
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>