[ose][media][acg] Migrate web media to ACG
:Release Notes:
LS2 migration on ACG for web media

:Detailed Notes:
Support a new security model(ACG) for web media except umedia client.
After umedia client is ready to provide ACG, we will apply them. For the
details refer to Issue below.
The rest of media backports except audio are ready to accept ACG but they
are not used in webos ose now. So, we are not test for them.

:Testing Performed:
Tested PLAT-65147 in RPI target.

:QA Notes:
Depends on https://gpro.lgsvl.com/#/c/215757
Depends on https://gpro.lgsvl.com/#/c/215017

Enact browser has not been applied for ACG yet. So, this patch does not
take account of playing youtube video on enact browser.

:Issues Addressed:
[PLAT-65140] Migrate chromium to ACG
[PLAT-65147] Verify that LS2 services for wam and web application are
             migrated to LS2 ACG

Change-Id: I12582711e19f4fbf87658a032ac0b46074be4a70
Reviewed-on: http://gpro.lgsvl.com/215773
Reviewed-by: Andre Rosa <andre.rosa@lge.com>
Tested-by: Andre Rosa <andre.rosa@lge.com>