[open][media][mse] Fix mute is not working in paused state
:Release Notes:
Fix audio mute is not working when playback is in paused state

:Detailed Notes:
We were sending the volume obtained from HTMLMediaElement to media
apis wrapper through PipelineImpl::RendererWrapper::SetVolume.
But it check the playing status before sending the volume level to
RendererImpl and further.
As a result media apis wrapper was not getting the volume level
as zero, being muted while in paused state.
In this patch set we changed code to provide the volume level info
directly to media apis wrapper.
To complete fix the below issue additional fix from audiod is
needed.
Currently mute is not effective even though we are passing zero
(0) to audiod

:Testing Performed:
1. Launch enact web browser
2. load www.youtube.com in a tab
3. play some video
4. after some time of play pause the playback
5. click the volume(speaker) icon on the playback control bar to
   mute the volume
6. resume play by clicking play button
7. audio should not be heard

:QA Notes:

:Issues Addressed:
[PLAT-65797] [webOS OSE] : Youtube Video Playback sound is heard,
             When volume is in mute.

Change-Id: Ic5afb27cbce6718f80d475f9f0e81d43642664b3
Reviewed-on: http://gpro.lgsvl.com/216119
Reviewed-by: Andre Rosa <andre.rosa@lge.com>
Tested-by: Andre Rosa <andre.rosa@lge.com>