[open][media][mse] Fix force unload on suspend media call
:Release Notes:
Fix video overlap on two tabs by doing force unload on suspend

:Detailed Notes:
When we play two videos in two tabs and switch video then
suspend is invoked for the background tab.
However the video is still seen in suspended state in the new
tab that is in foreground because the video plane in the
background tab is still on the screen and is seen through the
punch hole. Since there is no way to send the video plane to
background so as decided, we have to unload the video and
g-streamer pipeline
In this patch set we are unloading the media-player client which
will unload the pipeline internally.
And in resume the mediaplayer will be re-created since the inst
was destroyed already.
Force unload UMS videos are in https://gpro.lgsvl.com/#/c/214025/

Solved Conflicts:
  src/media/blink/webos/webmediaplayer_mse.cc
  src/media/blink/webos/webmediaplayer_mse.h
  src/media/webos/base/media_apis_wrapper.h

:Testing Performed:
1. Launch Enact Web browser
2. Load two videos in two tabs (any combination of URL and MSE)
3. Let the videos play and switch the tabs to see there is no
   overlap of videos from the background tab in foreground tab.

:QA Notes:

:Issues Addressed:
[PLAT-62988] [webOS OSE] Two videos of other tab are playback
             overlapped

Change-Id: I06a600075b73006270d46a9eb2bb60b8eb2eb7c2
Reviewed-on: http://gpro.lgsvl.com/216118
Reviewed-by: Andre Rosa <andre.rosa@lge.com>
Tested-by: Andre Rosa <andre.rosa@lge.com>