[open][media]mse] Fix compilation error for MSE VTG impl
:Release Notes:
Fix compilation error in WebMediaPlayerMSEVTGImpl for ScaleWebRect

:Detailed Notes:
Fix compilation error in WebMediaPlayerMSEVTGImpl for ScaleWebRect
The name changed from scaleWebRect to ScaleWebRect not reflecting
in WebMediaPlayerMSEVTGImpl.
Chnage to ScaleWebRect.

:Testing Performed:
Tested compilation error fix using GPVB build.

:QA Notes:
In OSE build this file is not used yet. But better to have the
fix in to the repo.

:Issues Addressed:
None

Change-Id: I99fbaf846cc84208d5a60dad3ca88ea93847da0e
Reviewed-on: http://gpro.lgsvl.com/216291
Reviewed-by: Biju Malayil <biju.malayil@lge.com>
Reviewed-by: Andre Rosa <andre.rosa@lge.com>
Tested-by: Andre Rosa <andre.rosa@lge.com>