[ose][media][acg] Make wam and app-shell to a container type for ACG
:Release Notes:
Make wam and app-shell to a container type and each application/page can
be registered as a application service in ls-hubd.

:Detailed Notes:
According to ACG model, wam and web applications are relevant
container-applications type. So, this patch follows the rules;
wam and app-shell which are browsers in chromium are registered as
a service of a container type at ls-bus. Whereas web applications or web
pages are registered as an application service type.

Developers provide different role files for these applications by using
"requiredPermissions" property in appinfo.json.

:Testing Performed:
Tested below the issue in RPI target.

:QA Notes:
Depends on https://gpro.lgsvl.com/220168
Depends on https://gpro.lgsvl.com/219711
Test-cases:QO-3159

:Issues Addressed:
[PLAT-67708] [OSE][ACG][App-shell] Implement Client side of Luna Service
as an application container

Change-Id: I99e4fb5017c9adca17b27067fd20d2888f7e4747
Reviewed-on: http://gpro.lgsvl.com/220254
Reviewed-by: Jeongran Jang <jeongran.jang@lge.com>
Tested-by: Jeongran Jang <jeongran.jang@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>
Reviewed-by: Andre Rosa <andre.rosa@lge.com>
Tested-by: Andre Rosa <andre.rosa@lge.com>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>